v5.5.0

+++ pydbtools/_wrangler.py
@@ -7,11 +7,11 @@ import logging
 import pprint
 import pandas as pd
 import re
-from typing import Iterator, Optional
-from urllib.parse import urlparse, urljoin, urlunparse
+from typing import Iterator, Optional, List
 import time
 import inspect
 import functools
+from arrow_pd_parser import reader
 
 from pydbtools.utils import (
     get_user_id_and_table_dir,
@@ -22,6 +22,7 @@ from pydbtools.utils import (
     get_boto_session,
     replace_temp_database_name_reference,
     _set_region_name,
+    s3_path_join,
 )
 
 
@@ -106,16 +107,21 @@ def init_athena_params(func=None, *, allow_boto3_session=False):  # noqa: C901
         # Set ctas_approach to True if not set.
         # Although awswrangler does this by default, we want to ensure
         # that timestamps are read in correctly to pandas using pyarrow.
-        # Therefore forcing the default option to be True incase future versions
-        # of wrangler change their default behaviour.
-        if "ctas_approach" in sig.parameters and argmap.get("ctas_approach") is None:
+        # Therefore forcing the default option to be True in case future
+        # versions of wrangler change their default behaviour.
+        if (
+            "ctas_approach" in sig.parameters
+            and argmap.get("ctas_approach") is None
+        ):
             argmap["ctas_approach"] = True
 
         # Set database to None or set to keyword temp when not needed
         if database_flag:
             if "ctas_approach" in sig.parameters and argmap["ctas_approach"]:
                 argmap["database"] = temp_db_name
-                _ = _create_temp_database(temp_db_name, boto3_session=boto3_session)
+                _ = _create_temp_database(
+                    temp_db_name, boto3_session=boto3_session
+                )
             elif argmap.get("database", "").lower() == "__temp__":
                 argmap["database"] = temp_db_name
             else:
@@ -132,7 +138,9 @@ def init_athena_params(func=None, *, allow_boto3_session=False):  # noqa: C901
             and "database" in sig.parameters
             and argmap.get("database") is None
         ):
-            argmap["database"] = get_database_name_from_sql(argmap.get("sql", ""))
+            argmap["database"] = get_database_name_from_sql(
+                argmap.get("sql", "")
+            )
 
         # Set pyarrow_additional_kwargs
         if (
@@ -165,6 +173,9 @@ start_query_execution = init_athena_params(ath.start_query_execution)
 stop_query_execution = init_athena_params(ath.stop_query_execution)
 wait_query = init_athena_params(ath.wait_query)
 tables = init_athena_params(wr.catalog.tables)
+create_ctas_table = init_athena_params(
+    ath.create_ctas_table, allow_boto3_session=True
+)
 
 
 @init_athena_params
@@ -180,7 +191,9 @@ def start_query_execution_and_wait(sql, *args, **kwargs):
     # to call the original unwrapped athena fun to ensure the wrapper fun
     # is not called again
     query_execution_id = ath.start_query_execution(sql, *args, **kwargs)
-    return ath.wait_query(query_execution_id, boto3_session=kwargs.get("boto3_session"))
+    return ath.wait_query(
+        query_execution_id, boto3_session=kwargs.get("boto3_session")
+    )
 
 
 def check_sql(sql: str):
@@ -233,7 +246,9 @@ def _create_temp_database(
 
     create_db_query = f"CREATE DATABASE IF NOT EXISTS {temp_db_name}"
 
-    q_e_id = ath.start_query_execution(create_db_query, boto3_session=boto3_session)
+    q_e_id = ath.start_query_execution(
+        create_db_query, boto3_session=boto3_session
+    )
     return ath.wait_query(q_e_id, boto3_session=boto3_session)
 
 
@@ -246,7 +261,7 @@ def create_temp_table(
     region_name: str = None,
 ):
     """
-    Create a table inside the database from create database
+    Create a table inside the temporary database from create table
 
     Args:
         sql (str):
@@ -258,14 +273,16 @@ def create_temp_table(
 
         force_ec2 (bool, optional):
             Boolean specifying if the user wants to force boto to get the
-            credentials from the EC2. This is for dbtools which is the R wrapper that
-            calls this package via reticulate and requires credentials to be refreshed
-            via the EC2 instance (and therefore sets this to True) - this is not
+            credentials from the EC2. This is for dbtools which is the R
+            wrapper that calls this package via reticulate and requires
+            credentials to be refreshed via the EC2 instance (and
+            therefore sets this to True) - this is not
             necessary when using this in Python. Default is False.
 
         region_name (str, optional):
             Name of the AWS region you want to run queries on. Defaults to
-            pydbtools.utils.aws_default_region (which if left unset is "eu-west-1").
+            pydbtools.utils.aws_default_region (which if left unset is
+            "eu-west-1").
     """
     region_name = _set_region_name(region_name)
     check_sql(sql)
@@ -287,7 +304,9 @@ def create_temp_table(
 
     drop_table_query = f"DROP TABLE IF EXISTS {temp_db_name}.{table_name}"
 
-    q_e_id = ath.start_query_execution(drop_table_query, boto3_session=boto3_session)
+    q_e_id = ath.start_query_execution(
+        drop_table_query, boto3_session=boto3_session
+    )
 
     _ = ath.wait_query(q_e_id, boto3_session=boto3_session)
 
@@ -304,6 +323,36 @@ def create_temp_table(
     q_e_id = ath.start_query_execution(ctas_query, boto3_session=boto3_session)
 
     ath.wait_query(q_e_id, boto3_session=boto3_session)
+
+
+def create_table(
+    sql: str,
+    database: str,
+    table: str,
+    location: str,
+    partition_cols: Optional[List[str]] = None,
+    boto3_session=None,
+):
+    """
+    Create a table in a database from a SELECT statement
+
+    Args:
+        sql (str): SQL starting with a WITH or SELECT clause
+        database (str): Database name
+        table (str): Table name
+        location (str): S3 path to where the table should be stored
+        partition_cols (List[str]): partition columns (optional)
+        boto3_session: optional boto3 session
+    """
+    return ath.create_ctas_table(
+        sql=sql,
+        ctas_database=database,
+        ctas_table=table,
+        s3_output=s3_path_join(location, table + ".parquet"),
+        partitioning_info=partition_cols,
+        wait=True,
+        boto3_session=boto3_session,
+    )
 
 
 def _create_temp_table_in_sql(sql: str) -> bool:
@@ -433,15 +482,23 @@ def delete_table_and_data(table: str, database: str, boto3_session=None):
     Args:
         table (str): The table name to drop.
         database (str): The database name.
+
+    Returns:
+        True if table exists and is deleted, False if table
+        does not exist
     """
 
-    path = wr.catalog.get_table_location(
-        database=database, table=table, boto3_session=boto3_session
-    )
-    wr.s3.delete_objects(path, boto3_session=boto3_session)
-    wr.catalog.delete_table_if_exists(
-        database=database, table=table, boto3_session=boto3_session
-    )
+    if table in list(tables(database=database, limit=None)["Table"]):
+        path = wr.catalog.get_table_location(
+            database=database, table=table, boto3_session=boto3_session
+        )
+        wr.s3.delete_objects(path, boto3_session=boto3_session)
+        wr.catalog.delete_table_if_exists(
+            database=database, table=table, boto3_session=boto3_session
+        )
+        return True
+    else:
+        return False
 
 
 @init_athena_params(allow_boto3_session=True)
@@ -451,11 +508,21 @@ def delete_database_and_data(database: str, boto3_session=None):
 
     Args:
         database (str): The database name to drop.
-    """
 
-    for table in wr.catalog.get_tables(database=database, boto3_session=boto3_session):
-        delete_table_and_data(table["Name"], database, boto3_session=boto3_session)
+    Returns:
+        True if database exists and is deleted, False if database
+        does not exist
+    """
+    if database not in (db["Name"] for db in wr.catalog.get_databases()):
+        return False
+    for table in wr.catalog.get_tables(
+        database=database, boto3_session=boto3_session
+    ):
+        delete_table_and_data(
+            table["Name"], database, boto3_session=boto3_session
+        )
     wr.catalog.delete_database(database, boto3_session=boto3_session)
+    return True
 
 
 @init_athena_params(allow_boto3_session=True)
@@ -505,7 +572,10 @@ def save_query_to_parquet(sql: str, file_path: str) -> None:
         file_path (str): The path to save the result to.
 
     Examples:
-    save_query_to_parquet("select * from my database.my_table", "result.parquet")
+    save_query_to_parquet(
+        "select * from my database.my_table",
+        "result.parquet"
+    )
     """
 
     df = read_sql_query(sql)
@@ -514,46 +584,137 @@ def save_query_to_parquet(sql: str, file_path: str) -> None:
     return None
 
 
-def s3_path_join(base: str, url: str, allow_fragments=True) -> str:
-    """
-    Joins a base S3 path and a URL. Acts the same as urllib.parse.urljoin,
-    which doesn't work for S3 paths.
-
-    Args:
-        base (str): Base S3 URL
-        url (str):
-    """
-    p = urlparse(base)
-    return urlunparse(p._replace(path=urljoin(p.path, url, allow_fragments)))
-
-
 @init_athena_params(allow_boto3_session=True)
-def dataframe_to_temp_table(df: pd.DataFrame, table: str, boto3_session=None) -> None:
+def dataframe_to_temp_table(
+    df: pd.DataFrame, table: str, boto3_session=None
+) -> None:
     """
     Creates a temporary table from a dataframe.
 
     Args:
         df (pandas.DataFrame): A pandas DataFrame
         table (str): The name of the table in the temporary database
+        boto3_session: opeional boto3 sesssion
     """
     user_id, table_dir = get_user_id_and_table_dir()
     db = get_database_name_from_userid(user_id)
     _create_temp_database(db)
-    # Clean up existing table if necessary
-    wr.catalog.delete_table_if_exists(
-        database=db, table=table, boto3_session=boto3_session
-    )
-    # Write table
-    # - Include timestamp in path to avoid permissions problems with
-    #   previous sessions
+
+    # Include timestamp in path to avoid permissions problems with
+    # previous sessions
     ts = str(time.time()).replace(".", "")
-    path = s3_path_join(table_dir, f"{ts}/{table}.parquet")
+    path = s3_path_join(table_dir, ts, table)
+    dataframe_to_table(df, db, table, path, boto3_session=boto3_session)
+
+
+@init_athena_params(allow_boto3_session=True)
+def dataframe_to_table(
+    df: pd.DataFrame,
+    database: str,
+    table: str,
+    location: str,
+    mode: str = "overwrite",
+    partition_cols: Optional[List[str]] = None,
+    boto3_session=None,
+    **kwargs,
+) -> None:
+    """
+    Creates a table from a dataframe.
+
+    Args:
+        df (pandas.DataFrame): A pandas DataFrame
+        database (str): Database name
+        table (str): Table name
+        location (str): S3 path to where the table should be stored
+        mode (str): "overwrite" (default), "append", or "overwrite_partitions"
+        partition_cols (List[str]): partition columns (optional)
+        boto3_session: optional boto3 session
+        **kwargs: arguments for to_parquet
+    """
+
+    # Write table
     wr.s3.to_parquet(
         df,
-        path=path,
+        path=s3_path_join(location, table + ".parquet"),
         dataset=True,
-        database=db,
+        database=database,
         table=table,
         boto3_session=boto3_session,
-        mode="overwrite",
+        mode=mode,
+        partition_cols=partition_cols,
+        compression="snappy",
+        **kwargs,
     )
+
+
+def create_database(database: str, **kwargs) -> bool:
+    """
+    Creates a new database.
+
+    Args:
+        database (str): The name of the database
+
+    Returns:
+        False if the database already exists, True if
+        it has been created.
+    """
+
+    if database in (db["Name"] for db in wr.catalog.get_databases()):
+        return False
+    wr.catalog.create_database(database, **kwargs)
+    return True
+
+
+@init_athena_params(allow_boto3_session=True)
+def file_to_table(
+    path: str,
+    database: str,
+    table: str,
+    location: str,
+    mode: str = "overwrite",
+    partition_cols: Optional[List[str]] = None,
+    boto3_session=None,
+    chunksize=None,
+    metadata=None,
+    **kwargs,
+) -> None:
+    """
+    Writes a csv, json, or parquet file to a database table.
+
+    Args:
+        path (str): The location of the file
+        database (str): database name
+        table (str): table name
+        location (str): s3 file path to table
+        mode (str): "overwrite" (default), "append", "overwrite_partitions"
+        partition_cols (List[str]): partition columns (optional)
+        boto3_session: optional boto3 session
+        chunksize Union[int,str]: size of chunks in memory or rows,
+            e.g. "100MB", 100000
+        metadata: mojap_metadata instance
+        **kwargs: arguments for arrow_pd_parser.reader.read
+            e.g. use chunksize for very large files, metadata
+            to apply metadata
+    """
+
+    dfs = reader.read(path, chunksize=chunksize, metadata=metadata, **kwargs)
+    if isinstance(dfs, pd.DataFrame):
+        # Convert single dataframe to iterator
+        dfs = iter([dfs])
+    elif mode == "overwrite_partitions":
+        raise ValueError(
+            "overwrite_partitions and a set chunksize "
+            + "can't be used at the same time"
+        )
+
+    for df in dfs:
+        dataframe_to_table(
+            df,
+            database,
+            table,
+            location,
+            partition_cols=partition_cols,
+            mode=mode,
+            boto3_session=boto3_session,
+        )
+        mode = "append"
